backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
python-docx==1.1.0
python-pptx==0.6.23
openai==1.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator>=2.0.0

